refactor: add UserProfile components
