+++ src/components/organisms/text/Text.tsx
@@ -2,7 +2,7 @@ import { format, parseISO } from 'date-fns';
 import type TextType from '../../../lib/models/Text';
 import UserAvatar from '../../atoms/UserAvatar';
 import UserLink from '../../atoms/UserLink';
-import UserName from '../UserName';
+import UserName from '../user/UserName';
 
 type Props = {
   text: TextType;
